+++ src/main/java/ru/vyarus/guice/persist/orient/repository/command/script/Script.java
@@ -20,6 +20,10 @@ import java.util.Collection;
  * {@link ru.vyarus.guice.persist.orient.repository.command.ext.elvar.ElVar}).</p>
  * <p>By default, sql function expected, but if other language required set it explicitly.</p>
  * <p>Scripts are very similar to {@link ru.vyarus.guice.persist.orient.repository.command.function.Function}</p>
+ * <p>WARNING: script did not support positional parameters (?)! Use named parameters or reference positional
+ * parameters by name (e.g. ':0' for first parameter instead of '?')</p>
+ * <p>Note that using script instead of update query, like this "begin update... commit" solves
+ * concurrent modification problem, even without retry.</p>
  *
  * @author Vyacheslav Rusakov
  * @see <a href="http://www.orientechnologies.com/docs/last/orientdb.wiki/SQL-batch.html">docs</a>
@@ -43,6 +47,7 @@ public @interface Script {
     /**
      * Use this clause to specify a collection impl to autobox result lists into. The impl must
      * have a default no-arg constructor and be a subclass of {@code java.util.Collection}.
+     *
      * @return configured autoboxing collection class.
      */
     Class<? extends Collection> returnAs() default Collection.class;