+++ src/test/groovy/ru/vyarus/guice/persist/orient/repository/command/script/ScriptCases.groovy
@@ -16,6 +16,16 @@ interface ScriptCases {
             return \$model.nick""")
     String nick()
 
+    // positional parameters not supported, but they could be referenced as named (':0')
+    @Script("""
+        update model set name = ?
+        """)
+    void positional(String name)
+
+    @Script("""
+        update model set name = :0
+        """)
+    void positionalHack(String name)
 
     @Script("""
             begin